Second Commit
Got the Search Hijacking algorithm to work! Turns out I was looking in the wrong place, I was trying to change the state twice (via refreshing the component) when all I needed to do was to update the props, which would have passed the updated search term to the other methods!

Changed the layout by putting the search button next to the inputbar and  by adding a white background to the video list.